+++ apps/api/src/bulletinboard/bulletinboard.controller.ts
@@ -20,7 +20,7 @@ import APPS from '@libs/appconfig/constants/apps';
 import BULLETIN_ATTACHMENTS_PATH from '@libs/bulletinBoard/constants/bulletinAttachmentsPaths';
 import { RequestResponseContentType } from '@libs/common/types/http-methods';
 import BulletinBoardService from './bulletinboard.service';
-import GetCurrentUser from '../common/decorators/getUser.decorator';
+import GetCurrentUser from '../common/decorators/getCurrentUser.decorator';
 import GetToken from '../common/decorators/getToken.decorator';
 import { checkAttachmentFile, createAttachmentUploadOptions } from '../filesystem/multer.utilities';
